Remove SbAlertsService from the SbAlertComponent to allow singleton
The Service has been moved from the constructor to an input. This allows for a parent module to pass on the service as an input and maintain the singleton pattern of the service.
Important to mention: The service needs to implement the singleton pattern, because one should be able to open a alert from anywhere in the parent module.

Also the SbAlertNotFoundError is implemented in the openAlert method of the service. The error gets thrown when the specified alert is not found.